fix(Marketplace): use separate dialog component for adding apps
