+++ dashboard/src2/components/NewAppDialog.vue
@@ -7,9 +7,8 @@
 				{
 					label: 'Add App',
 					variant: 'solid',
+					disabled: !app || !appValidated,
 					onClick() {
-						app.version = selectedVersion.value || options.versions[0].name;
-						app.branch = selectedBranch.value;
 						$emit('app-added', app);
 						show = false;
 					}
@@ -30,7 +29,7 @@
 				v-model="tabIndex"
 				v-slot="{ tab }"
 			>
-				<div class="-ml-0.5 pl-1">
+				<div class="-ml-0.5 p-1">
 					<div v-if="tab.value === 'public-github-app'" class="space-y-4">
 						<div class="mt-4 flex items-end space-x-2">
 							<FormControl
@@ -66,120 +65,12 @@
 						</div>
 					</div>
 					<div v-else-if="tab.value === 'your-github-app'" class="pt-4">
-						<div
-							v-if="$resources.options.loading"
-							class="mt-2 flex justify-center"
-						>
-							<LoadingText />
-						</div>
-						<div
-							class="flex justify-center pt-2"
-							v-else-if="!options?.authorized"
-						>
-							<Button
-								v-if="requiresReAuth"
-								variant="solid"
-								icon-left="github"
-								label="Re-authorize GitHub"
-								@click="$resources.clearAccessToken.submit()"
-								:loading="$resources.clearAccessToken.loading"
-							/>
-							<Button
-								v-if="needsAuthorization"
-								variant="solid"
-								icon-left="github"
-								label="Connect To GitHub"
-								:link="options.installation_url + '?state=' + state"
-							/>
-						</div>
-						<div v-else class="space-y-4">
-							<FormControl
-								type="autocomplete"
-								label="Choose GitHub User / Organization"
-								:options="
-									options.installations.map(i => ({
-										label: i.login,
-										value: i
-									}))
-								"
-								v-model="selectedGithubUser"
-							>
-								<template #prefix>
-									<img
-										v-if="selectedGithubUser"
-										:src="selectedGithubUser?.value?.image"
-										class="mr-2 h-4 w-4 rounded-full"
-									/>
-									<FeatherIcon v-else name="users" class="mr-2 h-4 w-4" />
-								</template>
-								<template #item-prefix="{ active, selected, option }">
-									<img
-										v-if="option.value?.image"
-										:src="option.value.image"
-										class="mr-2 h-4 w-4 rounded-full"
-									/>
-									<FeatherIcon v-else name="user" class="mr-2 h-4 w-4" />
-								</template>
-							</FormControl>
-							<span class="text-sm text-gray-600">
-								Don't see your organization?
-								<Link
-									:href="options.installation_url + '?state=' + state"
-									class="font-medium"
-								>
-									Add from GitHub
-								</Link>
-							</span>
-							<FormControl
-								type="autocomplete"
-								v-if="selectedGithubUser"
-								label="Choose GitHub Repository"
-								:options="
-									selectedGithubUser.value.repos.map(r => ({
-										label: r.name,
-										value: r.name
-									}))
-								"
-								v-model="selectedGithubRepository"
-							>
-								<template #prefix>
-									<FeatherIcon name="book" class="mr-2 h-4 w-4" />
-								</template>
-								<template #item-prefix="{ active, selected, option }">
-									<FeatherIcon
-										:name="option.value.private ? 'lock' : 'book'"
-										class="mr-2 h-4 w-4"
-									/>
-								</template>
-							</FormControl>
-
-							<p v-if="selectedGithubUser" class="!mt-2 text-sm text-gray-600">
-								Don't see your repository here?
-								<Link :href="selectedGithubUser.value.url" class="font-medium">
-									Add from GitHub
-								</Link>
-							</p>
-							<FormControl
-								v-if="selectedGithubRepository"
-								type="autocomplete"
-								label="Choose Branch"
-								:options="branchOptions"
-								v-model="selectedBranch"
-							>
-								<template #prefix>
-									<FeatherIcon name="git-branch" class="mr-2 h-4 w-4" />
-								</template>
-							</FormControl>
-						</div>
+						<GitHubAppSelector
+							v-model="app"
+							@validateApp="data => $resources.validateApp.submit(data)"
+							@fieldChange="appValidated = false"
+						/>
 					</div>
-					<FormControl
-						v-if="showVersionSelector && selectedBranch"
-						class="mt-4"
-						type="autocomplete"
-						label="Choose Version"
-						:options="options.versions.map(v => v.name)"
-						v-model="selectedVersion"
-					/>
 					<div class="mt-4 space-y-2">
 						<div
 							v-if="$resources.validateApp.loading && !appValidated"
@@ -188,9 +79,12 @@
 							<LoadingIndicator class="mr-2 w-4" />
 							Validating app...
 						</div>
-						<div v-if="appValidated" class="flex text-base text-gray-700">
+						<div
+							v-if="appValidated && app"
+							class="flex text-base text-gray-700"
+						>
 							<GreenCheckIcon class="mr-2 w-4" />
-							Found {{ this.app.title }} ({{ this.app.name }})
+							Found {{ app.title }} ({{ app.name }})
 						</div>
 					</div>
 				</div>
@@ -205,18 +99,14 @@
 <script>
 import { FormControl, Tabs } from 'frappe-ui';
 import { DashboardError } from '../utils/error';
+import GitHubAppSelector from './GitHubAppSelector.vue';
 
 export default {
 	name: 'NewAppDialog',
 	components: {
+		GitHubAppSelector,
 		FTabs: Tabs,
 		FormControl
-	},
-	props: {
-		showVersionSelector: {
-			type: Boolean,
-			default: false
-		}
 	},
 	emits: ['app-added'],
 	data() {
@@ -226,22 +116,19 @@ export default {
 			tabIndex: 0,
 			githubAppLink: '',
 			selectedBranch: '',
-			selectedVersion: '',
 			appValidated: false,
-			requiresReAuth: false,
 			selectedGithubUser: null,
 			selectedGithubRepository: null,
 			tabs: [
 				{
 					label: 'Public GitHub App',
-					value: 'public-github-app',
-					condition: () => !this.showVersionSelector
+					value: 'public-github-app'
 				},
 				{
 					label: 'Your GitHub App',
 					value: 'your-github-app'
 				}
-			].filter(tab => tab.condition?.() ?? true)
+			]
 		};
 	},
 	watch: {
@@ -258,25 +145,6 @@ export default {
 			this.selectedBranch = '';
 			this.appValidated = false;
 		},
-		selectedGithubUser() {
-			this.selectedBranch = '';
-			this.appValidated = false;
-		},
-		selectedGithubRepository(val) {
-			this.appValidated = false;
-			this.$resources.branches.submit({
-				owner: this.selectedGithubUser?.label,
-				name: val?.label,
-				installation: this.selectedGithubUser?.value.id
-			});
-
-			if (this.selectedGithubUser) {
-				let defaultBranch = this.selectedGithubUser.value.repos.find(
-					r => r.name === val.label
-				).default_branch;
-				this.selectedBranch = { label: defaultBranch, value: defaultBranch };
-			} else this.selectedBranch = '';
-		},
 		selectedBranch(newSelectedBranch) {
 			if (this.appOwner && this.appName && newSelectedBranch)
 				this.$resources.validateApp.submit({
@@ -288,17 +156,6 @@ export default {
 		}
 	},
 	resources: {
-		options() {
-			return {
-				url: 'press.api.github.options',
-				auto: true,
-				onError(error) {
-					if (error.messages.includes('Bad credentials')) {
-						this.requiresReAuth = true;
-					}
-				}
-			};
-		},
 		validateApp() {
 			return {
 				url: 'press.api.github.app',
@@ -307,11 +164,11 @@ export default {
 					if (!data) {
 						return;
 					}
-
 					let repository_url = this.githubAppLink;
 					if (!repository_url) {
-						var repo = this.selectedGithubRepository?.label || data.name;
-						repository_url = `https://github.com/${this.selectedGithubUser.label}/${repo}`;
+						const repo_owner = this.selectedGithubUser?.label || data.owner;
+						const repo = this.selectedGithubRepository?.label || data.name;
+						repository_url = `https://github.com/${repo_owner}/${repo}`;
 					}
 
 					this.app = {
@@ -319,7 +176,7 @@ export default {
 						title: data.title,
 						repository_url,
 						github_installation_id: this.selectedGithubUser?.value.id,
-						branch: this.selectedBranch.value
+						branch: this.selectedBranch.value || data.branch
 					};
 				}
 			};
@@ -332,7 +189,7 @@ export default {
 						/^(https?:\/\/)?(www\.)?github\.com\/([a-zA-Z0-9_.\-]+)\/([a-zA-Z0-9_.\-]+)(\/)?$/;
 					const isValidUrl = githubUrlRegex.test(this.githubAppLink);
 
-					if (this.tabIndex === 0 && !isValidUrl) {
+					if (!isValidUrl) {
 						throw new DashboardError('Please enter a valid github link');
 					}
 				},
@@ -344,31 +201,16 @@ export default {
 						};
 				}
 			};
-		},
-		clearAccessToken() {
-			return {
-				url: 'press.api.github.clear_token_and_get_installation_url',
-				onSuccess(installation_url) {
-					window.location.href = installation_url + '?state=' + this.state;
-				}
-			};
 		}
 	},
 	computed: {
-		options() {
-			return this.$resources.options.data;
-		},
 		appOwner() {
-			if (this.tabIndex === 1) {
-				return this.selectedGithubUser?.label;
-			} else {
+			if (this.tabIndex === 0) {
 				return this.githubAppLink.split('/')[3];
 			}
 		},
 		appName() {
-			if (this.tabIndex === 1) {
-				return this.selectedGithubRepository?.label;
-			} else {
+			if (this.tabIndex === 0) {
 				return this.githubAppLink.split('/')[4].replace('.git', '');
 			}
 		},
@@ -377,19 +219,6 @@ export default {
 				label: branch.name,
 				value: branch.name
 			}));
-		},
-		needsAuthorization() {
-			if (this.$resources.options.loading) return false;
-			return (
-				this.$resources.options.data &&
-				(!this.$resources.options.data.authorized ||
-					this.$resources.options.data.installations.length === 0)
-			);
-		},
-		state() {
-			let location = window.location.href;
-			let state = { team: this.$team.name, url: location };
-			return btoa(JSON.stringify(state));
 		}
 	}
 };
