+++ dashboard/src2/objects/marketplace.js
@@ -2,7 +2,6 @@ import { defineAsyncComponent, h } from 'vue';
 import { confirmDialog, icon, renderDialog } from '../utils/components';
 import GenericDialog from '../components/GenericDialog.vue';
 import ObjectList from '../components/ObjectList.vue';
-import NewAppDialog from '../components/NewAppDialog.vue';
 import ChangeAppBranchDialog from '../components/marketplace/ChangeAppBranchDialog.vue';
 import { toast } from 'vue-sonner';
 import router from '../router';
@@ -59,7 +58,7 @@ export default {
 				width: 1.0
 			}
 		],
-		primaryAction({ listResource: apps }) {
+		primaryAction() {
 			return {
 				label: 'New App',
 				variant: 'solid',
@@ -67,28 +66,11 @@ export default {
 					prefix: icon('plus')
 				},
 				onClick() {
-					renderDialog(
-						h(NewAppDialog, {
-							showVersionSelector: true,
-							onAppAdded(app) {
-								toast.promise(apps.insert.submit(app), {
-									loading: 'Adding new app...',
-									success: () => {
-										router.push({
-											name: 'Marketplace App Detail Listing',
-											params: { name: app.name }
-										});
-										return 'New app added';
-									},
-									error: e => {
-										return e.messages.length
-											? e.messages.join('\n')
-											: e.message;
-									}
-								});
-							}
-						})
+					const NewAppDialog = defineAsyncComponent(() =>
+						import('../components/marketplace/NewMarketplaceAppDialog.vue')
 					);
+
+					renderDialog(h(NewAppDialog));
 				}
 			};
 		}
